atualizando resolucoes

+++ src/aula01/Resolucao02Java.java
@@ -7,19 +7,19 @@ public class Resolucao02Java {
 	public static void main(String[] args) {
 		
 		Scanner leia = new Scanner(System.in);
-		int nota1, nota2, nota3, nota4;
+		float nota1, nota2, nota3, nota4;
 		
 		System.out.println("Digite a nota 1: ");
-		nota1 = leia.nextInt();
+		nota1 = leia.nextFloat();
 		
 		System.out.println("Digite a nota 2: ");
-		nota2 = leia.nextInt();
+		nota2 = leia.nextFloat();
 
 		System.out.println("Digite a nota 3: ");
-		nota3 = leia.nextInt();
+		nota3 = leia.nextFloat();
 		
 		System.out.println("Digite a nota 4: ");
-		nota4 = leia.nextInt();
+		nota4 = leia.nextFloat();
 		
 		System.out.println("Média final: " + (nota1 + nota2 + nota3 + nota4)/4);
 	}
