+++ src/aula01/Resolucao03Java.java
@@ -7,19 +7,19 @@ public class Resolucao03Java {
 	public static void main(String[] args) {
 		
 		Scanner leia = new Scanner(System.in);
-		int salarioBruto, adicionalNoturno, horasExtras, descontos;
+		float salarioBruto, adicionalNoturno, horasExtras, descontos;
 		
 		System.out.println("Digite o salário bruto: ");
-		salarioBruto = leia.nextInt();
+		salarioBruto = leia.nextFloat();
 		
 		System.out.println("Digite o adicional noturno: ");
-		adicionalNoturno = leia.nextInt();
+		adicionalNoturno = leia.nextFloat();
 
 		System.out.println("Digite as horas extras: ");
-		horasExtras = leia.nextInt();
+		horasExtras = leia.nextFloat();
 		
 		System.out.println("Digite os descontos: ");
-		descontos = leia.nextInt();
+		descontos = leia.nextFloat();
 		
 		System.out.println("Salário líquido: " + (salarioBruto + adicionalNoturno + (horasExtras * 5) - descontos));
